book_erp_v2/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlmodel==0.0.22
jinja2==3.1.5
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
reportlab==4.2.5
requests==2.32.3